Move variant example to LanguageVariantSystem.cs, remove temp tests
