+++ ExampleMod/Localization/TranslationsNeeded.txt
@@ -1,3 +1,3 @@
-en-US, 770/770, 100%, missing 0
-ru-RU, 6/770, 1%, missing 764
-zh-Hans, 2/770, 0%, missing 768
+en-US, 787/787, 100%, missing 0
+ru-RU, 6/787, 1%, missing 781
+zh-Hans, 2/787, 0%, missing 785
